requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv
openai>=1.0.0
